Move all abstract data, selector to types file
